# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "description": "도메인 에러코드 상수(단일 출처). 라우터/서비스는 ``ErrorCode.X``만 사용한다.",
        "enum": [
          "SLOT_CONFLICT",
          "EMAIL_TAKEN",
          "ROOM_LIMIT_REACHED",
          "ROOM_NOT_FOUND",
          "RESERVATION_NOT_FOUND",
          "NOTIFICATION_NOT_FOUND",
          "ACCOUNT_NOT_FOUND",
          "CANCEL_WINDOW_PASSED",
          "REJECT_WINDOW_PASSED",
          "RESERVATION_NOT_COMPLETED",
          "REVIEW_ALREADY_EXISTS",
          "REVIEW_NOT_FOUND",
          "REVIEW_REPLY_FORBIDDEN",
          "REVIEW_REPLY_ALREADY_EXISTS",
          "FORBIDDEN_ROLE",
          "UNAUTHENTICATED",
          "VALIDATION_ERROR",
          "GEOCODING_UNAVAILABLE",
          "LLM_PROVIDER_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "description": "표준 에러 본문의 ``detail`` 객체(OpenAPI 계약 노출용).",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "표준 에러 응답. 후속 라우터가 ``responses={409: {\"model\": ErrorResponse}}``로 참조.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "description": "로그인 요청(Story 1.8).\n\n``email``은 의도적으로 ``str``(EmailStr 아님)이다 — 로그인은 *검증*이 아니라 *대조*이므로,\n이메일 형식 위반까지 422가 아니라 401로 단일화한다(enumeration·형식 누출 차단, AC2).\n정규화(``.strip().lower()``)는 서비스가 수행한다. ``password``도 정책 validator 없음.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutRequest": {
        "description": "로그아웃 요청(AC4 — 멱등). RN=본문 / 웹=쿠키(미전송 시 라우터가 쿠키 폴백).",
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "LogoutRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "description": "재발급 요청(AC3). RN=본문 / 웹=쿠키(미전송 시 라우터가 쿠키 폴백).",
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "회원가입 요청. 형식·정책·역할 위반은 Pydantic 검증 → 1.5 핸들러가 422로 단일화.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "role": {
            "enum": [
              "booker",
              "provider"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "role"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "토큰 발급 응답(로그인·회전 공용, AC1·AC3). 와이어는 snake_case 유지(L286).\n\n``refresh_token``은 **원문**만 반환한다(해시는 DB 내부 — 절대 응답에 노출 금지).",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserPublic": {
        "description": "가입 성공 응답(사용자 리소스). ``password_hash`` 필드 없음 — 해시 비노출(AC1).",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserPublic",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "DeskNow 백엔드 API (Story 1.2 골격).",
    "title": "DeskNow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "이메일·비밀번호로 로그인한다 → 200 + 토큰 쌍(본문) + 인증 쿠키 set(AC1·AC2).\n\n잘못된 자격(미존재·틀린 비번·비활성)은 서비스가 401 ``UNAUTHENTICATED``로 단일화한다.",
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "로그아웃한다 → 204(AC4 — 멱등). refresh 해시 행을 삭제하고 인증 쿠키를 제거한다.\n\n토큰이 본문/쿠키 모두 없거나 이미 무효여도 204로 정상 종료한다(멱등).",
        "operationId": "auth_logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "현재 인증된 사용자를 반환한다 → 200 + UserPublic(인증 필요, AC1·AC5 실증).\n\n프론트 세션 복원에 사용 + ``get_current_principal``(헤더/쿠키 추출)을 실증한다.",
        "operationId": "auth_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "refresh 토큰으로 새 토큰 쌍을 회전 발급한다 → 200(AC3).\n\n토큰은 본문(RN) 또는 쿠키(웹)에서 추출한다. 만료/위조/회전·로그아웃됨은 401.",
        "operationId": "auth_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "이메일·비밀번호·역할로 가입한다 → 201 + UserPublic(해시 비노출).",
        "operationId": "auth_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}